% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_calc.R
\name{dfcv}
\alias{dfcv}
\title{make data frame of critical values sup.H and sup.absH}
\usage{
dfcv(ts, sup.H, sup.absH, alpha)
}
\description{
make data frame of critical values sup.H and sup.absH
}
\keyword{internal}

