% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_calc.R
\name{EqKern}
\alias{EqKern}
\title{Definitions of equivalent kernels, same as in RDHonest package}
\usage{
EqKern(kernel = "uniform", boundary = TRUE, order = 0)
}
\description{
Definitions of equivalent kernels, same as in RDHonest package
}
\keyword{internal}

