% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{nsw}
\alias{nsw}
\title{National Supported Work (NWS) Demonstration data}
\format{Data frame with 2,675 observations on 12 variables:
\describe{
\item{treatment}{treatment indicator}
\item{age}{Age in years}
\item{education}{Years of schooling}
\item{black}{}
\item{hispanic}{}
\item{married}{}
\item{nodegree}{No high school degree}
\item{re74}{Earnings in 1974}
\item{re75}{Earnings in 1975}
\item{re78}{Earnings in 1978}
\item{ue74}{Unemployed (i.e. zero earnings) in 1974}
\item{ue74}{Unemployed (i.e. zero earnings) in 1974}
}}
\source{
Rajeev Dehejia's website,
    \url{http://users.nber.org/~rdehejia/nswdata2.html}
}
\usage{
nsw
}
\description{
Dataset from Dehejia and Wahba (1999).
}
\keyword{datasets}

