% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extreme_value.R
\name{PlotEVSnoopingCV}
\alias{PlotEVSnoopingCV}
\title{Compare critical values based on Gaussian and extreme value approximation}
\usage{
PlotEVSnoopingCV()
}
\value{
figure comparing critical values
}
\description{
Compare critical values based on Gaussian and extreme value approximation
}
