% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{rhc}
\alias{rhc}
\title{Right-heart catheterization data}
\format{There are 5,735 observations on 54 variables. The treatment and
    outcome variables are
\describe{
\item{rhc}{Right Heart Catheterization (RHC, also called Swan-Ganz catheter)
         received within 24 hours after entering study.}
\item{survival}{Indicator for 30-day survival}
}
Socioeconomic variables:
\describe{
\item{age}{Age (years)}
\item{sex}{Female indicator}
\item{race}{White/Black/Other}
\item{edu}{Years of education}
\item{income}{Under $11k / $11-$25k / $25-$50k / > $50k}
\item{insurance}{No insurance / Private / Medicare / Medicaid / Private &
    Medicare / Medicare & Medicaid}
}
Disease categories:
\describe{
\item{cat1}{Primary disease category. ARF (acute respiratory failure) / COPD
          (chronic obsructive pulmonary disease) / CHF (congestive heart
          failure) / Cirrhosis / Coma (nontraumatic coma) / Colon Cancer
          (metastatic to the liver) / Lung Cancer (stage III or IV) / MOSF
          w/Malignancy (multi-organ system failure) / MOSF w/Sepsis}
\item{cat2}{Secondary disease category. Same categories as cat1, plus "None"}
\item{cancer}{None/Localized/Metastatic}
}
Categories of admission diagnosis:
\describe{
\item{ad_resp}{Respiratory}
\item{ad_card}{Cardiovascular}
\item{ad_neuro}{Neurological}
\item{ad_gastr}{Gastrointestinal}
\item{ad_renal}{Renal}
\item{ad_meta}{Metabolic}
\item{ad_hema}{Hematologic}
\item{ad_seps}{Sepsis}
\item{ad_trauma}{Trauma}
\item{ad_ortho}{Orthopedic}
}
Comorbidities:
\describe{
\item{hx_cardio}{Cardiovascular symptoms}
\item{hx_chf}{Congestive Heart Failure}
\item{hx_dement}{Dementia, Stroke or Cerebral Infarct, Parkinson's Disease}
\item{hx_psych}{Psychiatric History, Active Psychosis or Severe Depression}
\item{hx_cpd}{Chronic Pulmonary Disease, Severe Pulmonary Disease, Very Severe
            Pulmonary Disease}
\item{hx_renal}{Chronic Renal Disease, Chronic Hemodialysis or Peritoneal
   Dialysis}
\item{hx_liver}{Cirrhosis, Hepatic Failure}
\item{hx_gibleed}{Upper GI Bleeding.}
\item{hx_malig}{Solid Tumor, Metastatic Disease, Chronic Leukemia/Myeloma,
   AcuteLeukemia, Lymphoma}
\item{hx_immuno}{Immunosupperssion, Organ Transplant, HIV Positive, Diabetes}
\item{hx_ami}{Definite Myocardial Infarction}
\item{hx_transfer}{Transfer (> 24 Hours) from Another Hospital}
}
Physiological status on Day 1:
\describe{
\item{dnr1}{Do Not Resuscitate (DNR) status}
\item{surv2md1}{SUPPORT model estimate of 2-month survival probability}
\item{aps1}{APACHE III score (Acute Physiology and Chronic Health Evaluation),
    ignoring Coma}
\item{scoma1}{SUPPORT Glasgow Coma Score}
\item{wtkilo1}{weight in kilograms}
\item{wt0}{wtkilo==0}
\item{temp1}{Temperature (Celcius)}
\item{meanbp1}{Mean Blood Pressure (mm Hg)}
\item{resp1}{Respiratory rate (breaths/min)}
\item{hrt1}{Heart rate (beats/min)}
\item{pafi1}{PaO_2/(0.01*FiO_2)}
\item{paco21}{PaCO_2  (partial pressure of arterial carbon dioxide)}
\item{ph1}{Serum pH (arterial)}
\item{wbc1}{White Blood Cell (WBC) count}
\item{hema1}{Hematocrit}
\item{sod1}{Serum sodium}
\item{pot1}{Serum potassium}
\item{crea1}{Serum creatinine}
\item{bili1}{Bilirubin}
\item{alb1}{Albumin}
}
Other variables: \describe{\item{das2d3pc}{Duke Activity Status Index (DASI)
2 weeks prior ## "rhc"}}}
\source{
\url{http://biostat.mc.vanderbilt.edu/wiki/Main/DataSets}
}
\usage{
rhc
}
\description{
Dataset from Connors et al. (1996). This version labels and modifies the
variables so that they match those in Hirano and Imbens' analysis.
}
\keyword{datasets}
