% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_calc.R
\name{SnoopingCV}
\alias{SnoopingCV}
\title{Snooping-adjusted critical value}
\usage{
SnoopingCV(ratio, kernel, order, onesided = FALSE, coverage = FALSE,
  level = 0.95)
}
\arguments{
\item{ratio}{ratio of maximum to minimum bandwidth, number between 1 and 100}

\item{kernel}{TODO uniform, triangular, and epanechnikov, luniform, quniform
for local constant (Nadaraya-Watson), local linear or local quadratic}

\item{order}{TODO}

\item{onesided}{Logical specifying whether the critical value corresponds to
a one-sided confidence interval (\code{FALSE} by default)}

\item{coverage}{report coverage of unadjusted CIs instead?}

\item{level}{number specifying confidence level, \code{0.95} by default,
other choices are \code{0.9} and \code{0.99}.

The function looks up appropriate critical value form a table of pre-computed
values.}
}
\value{
critical value
}
\description{
Look up appropriate adjusted critical value or coverage of an unadjusted
confidence band
}
\examples{
## look up appropriate 99\% critical value for a regression
## discontinuity design using a triangular kernel, with ratio of maximum to
## minimum bandwidths equal to 6.2
SnoopingCV(6.2, 'ltriangular', onesided=FALSE, level=0.99)
}

