% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_near_boundary.R
\name{SnoopingCVNearBd}
\alias{SnoopingCVNearBd}
\title{Snooping-adjusted critical values for estimation near a boundary.}
\usage{
SnoopingCVNearBd(S, T, bwratios, kernel, order, db, ngr, alpha = c(0.1, 0.05,
  0.01))
}
\arguments{
\item{S}{number of draws of the Gaussian process \eqn{\hat{\mathbb{H}}(h)}}

\item{T}{number of draws from a normal distribution in each draw of the
Gaussian process}

\item{bwratios}{Bandwidth ratios of maximum to minimum bandwidth for whcih to
compute critical values}

\item{kernel}{Kernel function \eqn{k(u)} supported on [-1,1] that takes a
vector or a matrix as an argument \eqn{u}.}

\item{order}{Order of local linear regression}

\item{db}{Local distance to boundary, equal to \eqn{x_{0}/\underline{h}},
where \eqn{x_{0}} is point of interest.}

\item{ngr}{number of grid points at which to evaluate the Gaussian process}

\item{alpha}{A vector of values determining the confidence level
\eqn{1-\alpha} at which to compute critical values}
}
\description{
Calculate one- and two-sided critical values \eqn{c_{1-\alpha}(t;k)} for
values of \eqn{t} in \code{bwratios} based on evaluating the Gaussian process
\eqn{\hat{\mathbb{H}}(h)} at \code{ngr} values of \eqn{h} in the interval
\eqn{[1/t,1]}.
}
\keyword{internal}

