% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_calc.R
\name{SnoopingTablesGraphs}
\alias{SnoopingTablesGraphs}
\title{Tables and graphs of snooping-robust critical values}
\usage{
SnoopingTablesGraphs(cvs, maxratio = 10, bwr.print = c(seq(1, 2, by = 0.2),
  3:10, 20, 50, 99.9))
}
\arguments{
\item{cvs}{Data frame returned by \code{\link{DFSnoopingCV}}}

\item{maxratio}{maximum ratio of maximum to minimum bandwidth to plot}

\item{bwr.print}{values of bandwidth ratios to print in summary tables}
}
\description{
Print tables of snooping-adjusted critical values and return ggplot2 object
for plotting them
}

