% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_calc.R
\name{DFSnoopingCV}
\alias{DFSnoopingCV}
\title{Calculate snooping-robust critical values for common kernels}
\usage{
DFSnoopingCV(S, T, grid.length)
}
\arguments{
\item{S}{number of draws of the Gaussian process \eqn{\hat{\mathbb{H}}(h)}}

\item{T}{number of draws from a normal distribution in each draw of the
Gaussian process}

\item{grid.length}{length of grid on which to evaluate the Gaussian process
\eqn{\hat{\mathbb{H}}(h)}}
}
\description{
Calaculates a data frame of snooping-robust critical values for uniform,
triangular and epanechnikov kernels and local constant, local, and quadratic
regression in the interior and on the boundary.
}

