% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_calc.R
\name{GridSnoopingCV}
\alias{GridSnoopingCV}
\title{Snooping-adjusted critical values for bandwidth ratios on a grid.}
\usage{
GridSnoopingCV(S, T, ngr, kernel, alpha = c(0.1, 0.05, 0.01))
}
\arguments{
\item{S}{number of draws of the Gaussian process \eqn{\hat{\mathbb{H}}(h)}}

\item{T}{number of draws from a normal distribution in each draw of the
Gaussian process}

\item{ngr}{number of grid points on which to evaluate the Gaussian process
\eqn{\hat{\mathbb{H}}(s)}, or else a vector of grid points}

\item{kernel}{Kernel function \eqn{k(u)} supported on [-1,1] that takes a
vector or a matrix as an argument \eqn{u}.}

\item{alpha}{A vector of values determining the confidence level
\eqn{1-\alpha} at which to compute critical values}
}
\description{
Calculate one- and two-sided critical values \eqn{c_{1-\alpha}(t;k)} for
\code{ngr} values of \eqn{t=\overline{h}/\underline{h}} in between 1 and 100
based on evaluating the Gaussian process \eqn{\hat{\mathbb{H}}(h)} at the
same values.
}
\keyword{internal}
